# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: Mini Forum API
  version: "1.0.0"
  description: |
    REST API for the Mini Forum platform. All endpoints are prefixed with `/api/v1`.

    **Base URLs:**
    - Development: `http://localhost:5000/api/v1`
    - Production: set via `VITE_API_URL` (frontend/admin-client) or `FORUM_API_URL` (vibe-content)

    **Authentication:** JWT Bearer token. Obtain via `POST /auth/login`, pass as `Authorization: Bearer <token>`.

    **Response format:** All responses follow a common envelope:
    ```json
    { "success": true, "message": "...", "data": { ... } }
    ```
    Paginated responses include:
    ```json
    { "success": true, "message": "...", "data": [...], "pagination": { "page": 1, "limit": 20, "total": 100, "totalPages": 5 } }
    ```
    Error responses:
    ```json
    { "success": false, "message": "...", "errors": { "field": ["error detail"] } }
    ```

    **Naming convention:** All response fields are `camelCase`. Request bodies accept both `camelCase` and `snake_case`.
servers:
  - url: http://localhost:5000/api/v1
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          description: Response payload (varies by endpoint)
    PaginationMeta:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    ErrorResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          enum: [false]
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    UserSummary:
      type: object
      required: [id, username, displayName, role]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [MEMBER, MODERATOR, ADMIN, BOT]
    AuthTokens:
      type: object
      required: [accessToken, refreshToken]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
...
    Notification:
      type: object
      required: [id, type, userId, isRead, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          description: Notification type (e.g. NEW_COMMENT, NEW_VOTE, REPLY)
        userId:
          type: integer
        isRead:
          type: boolean
        isDeleted:
          type: boolean
        actorId:
          type: integer
          nullable: true
        actor:
          $ref: "#/components/schemas/UserSummary"
        postId:
          type: integer
          nullable: true
        commentId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
paths:
...
  /auth/login:
    post:
      summary: Login with email/username and password
      operationId: login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, password]
              properties:
                identifier:
                  type: string
                  description: Email or username
                password:
                  type: string
                  format: password
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [user, tokens]
                        properties:
                          user:
                            $ref: "#/components/schemas/UserSummary"
                          tokens:
                            $ref: "#/components/schemas/AuthTokens"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/refresh:
    post:
      summary: Refresh access token
      operationId: refreshToken
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          tokens:
                            $ref: "#/components/schemas/AuthTokens"
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/check-email:
    get:
      summary: Check if email is available
      operationId: checkEmail
      tags: [Auth]
      parameters:
        - in: query
          name: email
          schema:
            type: string
            format: email
          required: true
      responses:
        "200":
          description: Availability result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          available:
                            type: boolean
  /auth/check-username:
    get:
      summary: Check if username is available
      operationId: checkUsername
      tags: [Auth]
      parameters:
        - in: query
          name: username
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Availability result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          available:
                            type: boolean

  # ==================== Posts ====================
...
  /notifications:
    get:
      summary: Get user notifications
      operationId: getNotifications
      tags: [Notifications]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: unreadOnly
          schema:
            type: boolean
      responses:
        "200":
          description: Paginated notifications
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Notification"
                      pagination:
                        $ref: "#/components/schemas/PaginationMeta"
    delete:
      summary: Delete all notifications
      operationId: deleteAllNotifications
      tags: [Notifications]
      security:
        - bearerAuth: []
      responses:
        "204":
          description: All notifications deleted
  /notifications/unread-count:
    get:
      summary: Get unread notification count
      operationId: getUnreadCount
      tags: [Notifications]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Unread count
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          count:
                            type: integer
  /notifications/read-all:
    patch:
      summary: Mark all notifications as read
      operationId: markAllNotificationsRead
      tags: [Notifications]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All notifications marked as read
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          count:
                            type: integer
                            description: Number of notifications marked as read
  /notifications/{id}/read:
    patch:
      summary: Mark a notification as read
      operationId: markNotificationRead
      tags: [Notifications]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification marked as read
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Notification"
        "404":
          description: Notification not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /notifications/{id}:
    delete:
      summary: Delete a notification
      operationId: deleteNotification
      tags: [Notifications]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Notification deleted
        "404":
          description: Notification not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...